Migrate ItemsList to Styled Components (#347)
