+++ src/index.tsx
@@ -17,7 +17,7 @@ export { default as ItemData } from 'itemData'
 export { default as ItemInfo } from 'itemInfo'
 export { default as ItemRadio, ItemRadioStatus } from 'itemRadio/ItemRadio'
 export { default as ItemRadioGroup } from 'itemRadioGroup'
-export { default as ItemsList } from 'itemsList'
+export { default as ItemsList, ItemsListDivider } from 'itemsList'
 export { default as Itinerary } from 'itinerary'
 export { default as Loader } from 'loader'
 export { default as Menu } from 'menu'
