+++ src/itemsList/ItemsList.unit.tsx
@@ -1,9 +1,8 @@
 import React from 'react'
-import { shallow, mount } from 'enzyme'
 import renderer from 'react-test-renderer'
 
 import ItemChoice from 'itemChoice'
-import ItemsList from './index'
+import ItemsList, { ItemsListDivider } from './index'
 
 describe('ItemsList', () => {
   it('Should render an unordered list', () => {
@@ -36,7 +35,7 @@ describe('ItemsList', () => {
         <ItemsList>
           <ItemChoice label="Item 1" href="#1" />
           <ItemChoice label="Item 2" href="#2" />
-          <ItemsList.Divider />
+          <ItemsListDivider />
           <ItemChoice label="Item 3" href="#3" />
         </ItemsList>,
       )
