+++ src/itemsList/specifications/index.md
@@ -1,7 +1,7 @@
 # ItemsList
 
 ```js
-import { ItemsList, ItemChoice } from '@blablacar/ui-library'
+import { ItemsList, ItemChoice, ItemsListDivider } from '@blablacar/ui-library'
 ```
 
 ## Playground
@@ -27,7 +27,7 @@ import { ItemsList, ItemChoice } from '@blablacar/ui-library'
   <ItemChoice label="Item 1" href="#1" />
   <ItemChoice label="Item 2" href="#2" />
   <ItemChoice label="Item 3" href="#3" />
-  <ItemsList.Divider />
+  <ItemsListDivider />
   <ItemChoice label="Item 4" href="#4" />
 </ItemsList>
 ```
